[Refactor] : ProjectEnvironment 파일 분리
Git : #92
