+++ Plugins/ProjectEnvironment/ProjectDescriptionHelpers/Helpers/Module/uFeatureModule.swift
@@ -7,31 +7,6 @@
 
 import ProjectDescription
 
-public protocol Module {
-  static var prefixPathString: String { get }
-  static var projectName: String { get }
-  static var featureName: String { get }
-  static var projectDepedency: TargetDependency { get }
-}
-
-public extension Module {
-  static var projectName: String {
-    String(describing: Self.self)
-  }
-  
-  static var featureName: String {
-    projectName
-  }
-  
-  static var path: Path {
-    .relativeToRoot(Self.prefixPathString + "/\(Self.projectName)")
-  }
-  
-  static var projectDepedency: TargetDependency {
-    .project(target: Self.featureName, path: path)
-  }
-}
-
 public protocol uFeatureModule: Module {}
 
 public extension uFeatureModule {
