+++ Projects/Core/NetworkStorage/Testings/TargetTypeTestings/MockExtensionTargetType.swift
@@ -17,7 +17,8 @@ public enum MockExtensionTargetType {
 
 extension MockExtensionTargetType: TargetType {
   public var baseURL: String {
-    return "https://baseURL"
+//    return "https://baseURL"
+    return Bundle.main.object(forInfoDictionaryKey: "BASE_URL") as? String ?? ""
   }
   
   public var path: String {
