+++ Projects/Core/NetworkStorage/Tests/ExtensionTargetTypeTests.swift
@@ -17,6 +17,11 @@ final class ExtensionTargetTypeTests: XCTestCase {
     targetType = .requestPlain
   }
   
+  func testFetchBaseURL() {
+    let url = "https://miniiad.duckdns.org"
+    XCTAssertEqual(targetType.baseURL, url)
+  }
+  
   func testGetURLRequestTaskRequestPlain() {
     targetType = .requestPlain
